feat: extract state directive handlers
